UI Element added to Program.cs
Added UI that uses while loop to show the menu options (which is probably not needed for this exercise but whatever) Switch to use as option to select instead of loads of if and else if statements. It has Method added for testing purposes which you can delete when when we have rest of the program done.